+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The provider's `container_api_url` configuration has been removed, as the comput
 
 ### Added
 - resource `dns_record`: support `SVCB`, `HTTPS` and `TLSA` types;
+- resource `dns_record`: support [Null MX](https://www.rfc-editor.org/rfc/rfc7505.html) ([#74](https://github.com/BunnyWay/terraform-provider-bunnynet/pull/74));
 
 ### Fixed
 - resource `compute_container_app`: fix flaky "attribute was null" error;